# Optimize render flamegraph