--- a/pkg/query/flamegraph_flat_test.go
+++ b/pkg/query/flamegraph_flat_test.go
@@ -227,7 +227,7 @@ func testGenerateFlamegraphFromProfile(t *testing.T, l metastorepb.MetastoreServ
 	sp, err := parcacol.NewArrowToProfileConverter(tracer, l).SymbolizeNormalizedProfile(ctx, profiles[0])
 	require.NoError(t, err)
 
-	fg, err := GenerateFlamegraphTable(ctx, tracer, sp, nodeTrimFraction)
+	fg, err := GenerateFlamegraphTable(ctx, tracer, sp, nodeTrimFraction, newTableConverterPool())
 	require.NoError(t, err)
 
 	return fg