The render flamegraph function can spend a lot of time allocating and gc'ing buffers. This adds a sync.pool for the buffers for rendering.

```
thor@thors-MacBook-Pro query % benchstat before.txt after.txt
name                                     old time/op    new time/op    delta
_GenerateFlamegraphTable_FromProfile-10     760µs ± 5%     696µs ± 5%  -8.35%  (p=0.000 n=10+10)

name                                     old alloc/op   new alloc/op   delta
_GenerateFlamegraphTable_FromProfile-10    1.23MB ± 0%    1.15MB ± 0%  -6.68%  (p=0.000 n=10+9)

name                                     old allocs/op  new allocs/op  delta
_GenerateFlamegraphTable_FromProfile-10     11.7k ± 0%     11.6k ± 0%  -0.64%  (p=0.000 n=10+10)
```

It produces moderate savings